Extract SelectRepository component from GitHub trigger panel
