+++ internal-packages/workflow-designer-ui/src/editor/properties-panel/trigger-node-properties-panel/ui/index.ts
@@ -1,2 +1,1 @@
-export * from "./label";
-export * from "./select";
+export * from "./select-repository";
