--- internal-packages/workflow-designer-ui/src/editor/properties-panel/trigger-node-properties-panel/ui/label.tsx
@@ -1,11 +1,0 @@
-import clsx from "clsx/lite";
-import { Label as LabelPrimitive } from "radix-ui";
-
-export function Label({ className, ...props }: LabelPrimitive.LabelProps) {
-	return (
-		<LabelPrimitive.Root
-			className={clsx("text-black-400 text-[12px]", className)}
-			{...props}
-		/>
-	);
-}
